+++ scripts/winterm/test.ps1
@@ -292,7 +292,7 @@ function Test-ShellExperienceFoundations
     }
 
     $manifest = Import-PowerShellDataFile -LiteralPath $moduleManifest
-    if ($manifest.ModuleVersion -ne '1.2.2' -or
+    if ($manifest.ModuleVersion -ne '1.2.3' -or
         $manifest.PrivateData.PSData.Prerelease -ne '' -or
         $manifest.PowerShellVersion -ne '5.1')
     {